+++ src/screens/conference/betaAccess/betaAccess.jsx
@@ -3,15 +3,13 @@ import PropTypes from 'prop-types'
 
 import InputButton from 'components/form/inputButton'
 
-import './betaAccess.css'
+import styles from './betaAccess.module.css'
 
 const BetaAccess = ({ validateAccessKey, error }) => (
-  <div className="beta-access">
-    <h1 className="beta-access-title">Beta Access key needed</h1>
-    <p>
-      The organizer hall is in closed-beta access, you need an invite key to be able to access it.
-    </p>
-    <div className="beta-access-form">
+  <div className={styles.wrapper}>
+    <h1 className={styles.title}>Beta Access key needed</h1>
+    <p>The organizer hall is in closed-beta access, you need a key to access it.</p>
+    <div className={styles.form}>
       <InputButton
         btnLabel="Access beta"
         type="text"
@@ -19,7 +17,12 @@ const BetaAccess = ({ validateAccessKey, error }) => (
         onClick={validateAccessKey}
       />
     </div>
-    {error && <div className="beta-access-error">{error}</div>}
+    <div className={styles.request}>
+      <a href="https://forms.gle/AnArRCSHibmG59zw7" target="NEW">
+        Request a beta key
+      </a>
+    </div>
+    {error && <div className={styles.error}>{error}</div>}
   </div>
 )